@@ -1,6 +1,6 @@
 class Monologue::Admin::SessionsController < Monologue::Admin::BaseController
-  skip_before_filter :authenticate_user!
-  
+  skip_before_action :authenticate_user!
+
   def new
   end
 
