@@ -3,7 +3,7 @@ class Monologue::ApplicationController < ApplicationController
 
   layout Monologue::Config.layout if Monologue::Config.layout # TODO: find a way to test that. It was asked in issue #54 (https://github.com/jipiboily/monologue/issues/54)
 
-  before_filter :recent_posts, :all_tags, :archive_posts
+  before_action :recent_posts, :all_tags, :archive_posts
 
   def recent_posts
     @recent_posts = Monologue::Post.published.limit(3)
