@@ -1,4 +1,4 @@
-class CreateMonologuePosts < ActiveRecord::Migration
+class CreateMonologuePosts < ActiveRecord::Migration[5.1]
   def change
     create_table :monologue_posts do |t|
       t.integer :posts_revision_id
