@@ -1,4 +1,4 @@
-class JoinPostsAndTags < ActiveRecord::Migration
+class JoinPostsAndTags < ActiveRecord::Migration[5.1]
   def change
      create_table :posts_tags, id:false do |t|
        t.integer :post_id,:tag_id
