@@ -1,4 +1,4 @@
-class AddIndexToPostsRevisionUrl < ActiveRecord::Migration
+class AddIndexToPostsRevisionUrl < ActiveRecord::Migration[5.1]
   def change
     add_index :monologue_posts_revisions, :url
   end
