@@ -1,4 +1,4 @@
-class RenamePostTagsTable < ActiveRecord::Migration
+class RenamePostTagsTable < ActiveRecord::Migration[5.1]
   def change
     rename_table :posts_tags, :monologue_posts_tags
   end
