@@ -1,6 +1,6 @@
-# Read about factories at http://github.com/thoughtbot/factory_girl
+# Read about factories at http://github.com/thoughtbot/factory_bot
 
-FactoryGirl.define do
+FactoryBot.define do
   factory :user, class: Monologue::User do
     sequence(:name){|n| "test #{n}"}
     sequence(:email){ |n| "test#{n}@example.com"}
@@ -9,6 +9,6 @@ FactoryGirl.define do
   end
 
   factory :user_with_post, class: Monologue::User, parent: :user do |user|
-    user.after_create { |u| Factory(:post, user: u) }
+    after(:create) { |u| FactoryBot.build(:post, user: u) }
   end
 end
