@@ -1,8 +1,8 @@
 require 'spec_helper'
 
-describe Monologue::Tag do
+RSpec.describe Monologue::Tag do
   before(:each) do
-    @tag= Factory(:tag)
+    @tag= FactoryBot.build(:tag)
   end
 
   it "is valid with valid attributes" do
@@ -11,11 +11,12 @@ describe Monologue::Tag do
 
   describe "validations" do
     it "is not possible to have save another tag with the same name" do
-       expect { Factory(:tag) }.to raise_error(ActiveRecord::RecordInvalid)
+      FactoryBot.create(:tag, name: "foo")
+      expect { FactoryBot.create(:tag, name: "foo") }.to raise_error(ActiveRecord::RecordInvalid)
     end
 
     it "should require the name to be set" do
-      expect { Factory(:tag,name:nil) }.to raise_error(ActiveRecord::RecordInvalid)
+      expect { FactoryBot.create(:tag,name:nil) }.to raise_error(ActiveRecord::RecordInvalid)
     end
   end
 end
