@@ -1,5 +1,6 @@
 require 'spec_helper'
-describe "main_app_integration" do
+
+RSpec.describe "main_app_integration" do
   it "should use main_app layout" do
     # TODO: make a working test with that. Changing layout on the fly does not seem to work.
     # Monologue.layout = "layouts/application"
