@@ -1,8 +1,8 @@
 # encoding: UTF-8
 require 'spec_helper'
-describe "pagination" do
+RSpec.describe "pagination" do
   before(:each) do
-    22.times { |i| Factory(:post, title: "post #{i}") }
+    22.times { |i| FactoryBot.build(:post, title: "post #{i}") }
   end
 
   it "should not show all posts" do
