@@ -1,11 +1,11 @@
 require 'spec_helper'
-describe "posts" do
+RSpec.describe "posts" do
   before(:each) do
-    Factory(:post, title: "post X")
+    FactoryBot.build(:post, title: "post X")
   end
 
   it "lists posts" do
-    Factory(:post, title: "post Y")
+    FactoryBot.build(:post, title: "post Y")
     visit "/monologue"
 
     page.should have_content("post X")
@@ -29,13 +29,13 @@ describe "posts" do
   end
 
   it "should not show post with published date in the future" do
-    Factory(:post, published_at: DateTime.new(3000), title: "I am Marty McFly")
+    FactoryBot.build(:post, published_at: DateTime.new(3000), title: "I am Marty McFly")
     visit root_path
     page.should_not have_content "I am Marty McFly"
   end
 
   it "should not show an unpublished post" do
-    Factory(:post, published: false, title: "I am Marty McFly")
+    FactoryBot.build(:post, published: false, title: "I am Marty McFly")
     visit root_path
     page.should_not have_content "I am Marty McFly"
   end
