@@ -1,16 +1,17 @@
 require 'spec_helper'
-describe "preview" do
+
+RSpec.describe "preview", type: :request do
   before(:each) do
     url ="post/1"
     @post_path = "/monologue/#{url}"
     @post_title = "post 1"
-    @post = Factory(:post, title: @post_title, url: url)
+    @post = FactoryBot.build(:post, title: @post_title, url: url)
     ActionController::Base.perform_caching = true
   end
 
   it "verify unpublished posts are not public" do
     visit root_path
-    Factory(:unpublished_post)
+    FactoryBot.build(:unpublished_post)
     page.should_not have_content("unpublished")
     visit "/monologue/unpublished"
     page.should have_content("You may have mistyped the address or the page may have moved")
