@@ -3,8 +3,8 @@ describe "tag cloud" do
   describe "Viewing the tag cloud" do
 
     before(:each) do
-      Factory(:post_with_tags)
-      post = Factory(:post, title: "Future post", published_at: DateTime.new(3000))
+      FactoryBot.build(:post_with_tags)
+      post = FactoryBot.build(:post, title: "Future post", published_at: DateTime.new(3000))
       post.tag!(["Rails", "another tag"])
     end
 
