@@ -1,10 +1,10 @@
 # encoding: UTF-8
 
 require 'spec_helper'
-describe "tags" do
+RSpec.describe "tags", type: :feature do
   describe "Viewing the list of posts with tags" do
     before(:each) do
-      Factory(:post_with_tags, title: "post X")
+      FactoryBot.build(:post_with_tags, title: "post X")
     end
 
     it "should display the tags for the posts as a link" do
@@ -17,8 +17,8 @@ describe "tags" do
 
   describe "filtering by a given tag" do
     before(:each) do
-      @post = Factory(:post_with_tags, title: "post X")
-      Factory(:post, title: "post Z")
+      @post = FactoryBot.build(:post_with_tags, title: "post X")
+      FactoryBot.build(:post, title: "post Z")
     end
 
     it "should only display posts with the given tag" do
@@ -30,7 +30,7 @@ describe "tags" do
     end
 
     it "should not display posts with tags with future publication date" do
-      post = Factory(:post, title: "we need to reach 88 miles per hour", published_at: DateTime.new(3000))
+      post = FactoryBot.build(:post, title: "we need to reach 88 miles per hour", published_at: DateTime.new(3000))
       post.tag!(["rails","another tag"])
       visit "/monologue"
       click_on "Rails"
@@ -40,7 +40,7 @@ describe "tags" do
 
     it "should work with non-latin tag" do
       pending "temporarily disabled because of problem with ruby 1.9.3"
-      post = Factory(:post, title: "non-latin tag post title", published_at: DateTime.new(3000))
+      post = FactoryBot.build(:post, title: "non-latin tag post title", published_at: DateTime.new(3000))
       post.tag!(["rails","Тест"])
       visit "/monologue"
       click_on "Тест"
