@@ -1,7 +1,6 @@
 require 'capybara/rails'
 require 'capybara/dsl'
+
 RSpec.configure do |c|
-  c.include Capybara::DSL, example_group: {
-    file_path: /\bspec\/request\//
-  }
+  c.include Capybara::DSL, file_path: /\bspec\/request\//
 end
